% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edetect-package.R
\docType{package}
\name{edetect-package}
\alias{edetect}
\alias{edetect-package}
\title{edetect: Anytime-Valid Control Charts with E-Detectors}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Sequential change detection for process monitoring with e-detectors (Shin, Ramdas and Rinaldo, 2024, \doi{10.51387/23-NEJSDS51}): nonparametric, anytime-valid analogues of the CUSUM and Shiryaev-Roberts charts whose average run length to a false alarm is guaranteed for every distribution in a declared pre-change class (bounded, sub-Gaussian, Bernoulli or Poisson observations). Phase I/Phase II workflow, individual, p, c, u and regression charts, run-length simulation and auditable alarm reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://castlaboratory.github.io/edetect/}
  \item \url{https://github.com/castlaboratory/edetect}
  \item Report bugs at \url{https://github.com/castlaboratory/edetect/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})
  \item Raydonal Ospina \email{raydonal@castlab.org} (\href{https://orcid.org/0000-0002-9884-9090}{ORCID})
  \item Cristiano Ferraz \email{cferraz@castlab.org} (\href{https://orcid.org/0000-0002-6838-6734}{ORCID})
}

}
\keyword{internal}
