% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{edetect_init}
\alias{edetect_init}
\title{Initialise an e-detector}
\usage{
edetect_init(design)
}
\arguments{
\item{design}{An \code{\link[=edetect_design]{edetect_design()}}.}
}
\value{
An object of class \code{edetect_chart} holding the design, the running
state and an empty trajectory.
}
\description{
Initialise an e-detector
}
\examples{
design <- edetect_design(arl = 370, class = "bounded", center = 0.5, bounds = c(0, 1))
state <- edetect_init(design)
state
}
